Use Microsoft.Windows.CsWin32 for Win32 interop
Use the Microsoft.Windows.CsWin32 package to generate Win32 interop code
for the hexus daemon. This allows to have proper enums/constants for
the (documented) APIs such as CreateToolhelp32Snapshot, OpenProcess,
CreateRemoteThread, and WaitForSingleObject.

Also this allows to properly marshal the ProcessEntry32 structure as it
source generates an inline array instead of like it was before with DllImport
which used string marshalling.

CtrlRoutine is still manually defined as it is not documented and
WindowsCtrlType now includes all the signals that are documented to be
supported by a CtrlRoutine callback.