Add windows service hosting extension
If someone wants this they can use it, however `hexus startup` won't have it provably, as it requires the `secpol.msc` 'Log on as a service' policy to include the user account, the user account to have a password (if it is an MSA account then the microsoft account needs to have a password) and to provide it to the service to run.

The task scheduler does not have these requirements so it is better for most setups.